Time: 0 ms (100%), Space: 7.6 MB (96.36%) - LeetHub

+++ 0371-sum-of-two-integers/0371-sum-of-two-integers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int getSum(int a, int b) {
          while (b != 0) {
-            int carry = (unsigned)(a & b) << 1;
+            int carry = (a & b) << 1;
             a = a ^ b;
             b = carry;
         }
